# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Visca Corporation API",
        "description": "Dokumentasi Open Api Visca Corporation",
        "contact": {
            "email": "bimasaktiputra95@gmail.com"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Server Local"
        }
    ],
    "paths": {
...
        "/v1/produk/list": {
            "get": {
                "tags": [
                    "Produk"
                ],
                "summary": "List Data Produk Master",
                "description": "List Data Produk Master",
                "operationId": "List Data Produk Master",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Halaman page data untuk ditampilkan",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    },
                    {
                        "name": "category_id",
                        "in": "query",
                        "description": "Sort data berdasarkan id ketegori",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Tipe pencarian berdasarkan nama produk",
                        "required": false,
                        "schema": {
                            "type": "text"
                        },
                        "example": "Nasi Goreng Ayam"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List Data Produk Master",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "string",
                                                "example": [
                                                    "List Data Produk Master"
                                                ]
                                            },
                                            "collectionFormat": "multi"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/v1/produk/tambah-baru": {
            "post": {
                "tags": [
                    "Produk"
                ],
                "summary": "Tambah data produk master",
                "description": "Tambah data produk master",
                "operationId": "Tambah data produk master",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "category_id",
                                    "name",
                                    "modal",
                                    "price",
                                    "stock",
                                    "satuan",
                                    "is_jual"
                                ],
                                "properties": {
                                    "category_id": {
                                        "description": "Id kategori untuk produk yang akan ditambahkan",
                                        "type": "integer"
                                    },
                                    "name": {
                                        "description": "Nama Produk",
                                        "type": "text"
                                    },
                                    "modal": {
                                        "description": "Harga Beli Produk",
                                        "type": "integer"
                                    },
                                    "price": {
                                        "description": "Harga Jual Produk",
                                        "type": "integer"
                                    },
                                    "stock": {
                                        "description": "Jumlah stock produk yang akan ditambahkan",
                                        "type": "integer"
                                    },
                                    "satuan": {
                                        "description": "Satuan stock beli / jual",
                                        "type": "string",
                                        "enum": [
                                            "pcs",
                                            "bundle"
                                        ]
                                    },
                                    "is_jual": {
                                        "description": "Jika produk di jual (true) jika tidak (false)",
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Berhasil menambahkan data produk",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolval",
                                            "example": "true"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Berhasil menambahkan data produk"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Penambahan produk gagal, periksa kembali data validasi anda",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Penambahan produk gagal, cek kembali validasi anda"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/v1/produk/detail/{id}": {
            "get": {
                "tags": [
                    "Produk"
                ],
                "summary": "Detail data produk",
                "description": "Detail data produk",
                "operationId": "Detail data produk",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id Produk",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data detail produk berdasarkan id",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "string",
                                                "example": [
                                                    "Data detail produk berdasarkan id"
                                                ]
                                            },
                                            "collectionFormat": "multi"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Data detail produk tidak ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolval",
                                            "example": "false"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Data detail produk tidak ditemukan"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/v1/produk/update-data/{id}": {
            "put": {
                "tags": [
                    "Produk"
                ],
                "summary": "Update data produk",
                "description": "Update data produk",
                "operationId": "Update data produk",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id produk merchant",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "name",
                                    "modal",
                                    "price",
                                    "is_active",
                                    "is_jual"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Nama produk baru",
                                        "type": "text"
                                    },
                                    "modal": {
                                        "description": "Harga Beli baru",
                                        "type": "integer"
                                    },
                                    "price": {
                                        "description": "Harga Jual baru",
                                        "type": "integer"
                                    },
                                    "is_active": {
                                        "description": "Nonaktifkan produk",
                                        "type": "boolean"
                                    },
                                    "is_jual": {
                                        "description": "Tampilkan produk untuk di jual",
                                        "type": "text"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Berhasil update data produk",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolval",
                                            "example": "true"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Berhasil update data produk"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Data produk tidak ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolval",
                                            "example": "false"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Data produk tidak ditemukan"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}